Ignore blank lines, chunk messages up
Blank lines were being sent and received, which is a bit stupid and
contrary to what IRC does by default. This commit contains code which
prevents blank lines from being sent, and ignores any blank messages
that are received.

IRC seems to have a limit to the number of characters per message. This
was causing JSON payloads to be truncated and hence the receivers of
the message weren't able to decrypt. This code breaks the messages into
chunks of 200 chars and sends each chunk as a separate message to avoid
this problem. Down the track I might do some smarts around detecting
word boundaries, but this will have to do for now.